refac: 커뮤니티 페이지 리팩토링
