+++ src/frontend/src/components/atoms/Div/TabDivider.tsx
@@ -1,13 +1,8 @@
 import styled from "styled-components";
 
-const TabDivider = () => {
-  return <Divider />;
-};
-
-const Divider = styled.div`
+const TabDivider = styled.div`
   width: 100%;
   height: 0.0938rem;
-  /* margin-bottom: 0.5rem; */
   background-color: ${({ theme }) => theme.backgroundColor["user-tab"]};
 `;
