+++ src/frontend/src/components/molecules/Button/CommunityLogoUpload.stories.tsx
