+++ src/frontend/src/components/organisms/Modal/CommunitySettingModal.tsx
@@ -1,8 +1,8 @@
 import styled from "styled-components";
-import CommunitySettingBar from "../components/organisms/CommunitySettingBar";
-import CommunitySettingMember from "../components/organisms/CommunitySettingMember";
-import CommunitySettingDefault from "../components/organisms/CommunitySettingDefault";
-import CommunitySettingInvite from "../components/organisms/CommunitySettingInvite";
+import CommunitySettingBar from "../CommunitySettingBar";
+import CommunitySettingMember from "../CommunitySettingMember";
+import CommunitySettingDefault from "../CommunitySettingDefault";
+import CommunitySettingInvite from "../CommunitySettingInvite";
 import BackgroundModal from "@components/organisms/BackgroundModal";
 import CancelIcon from "@components/atoms/Icons/CancelIcon";
 import useCommunityStore, { SettingStatusType } from "@store/useCommunityStore";
@@ -19,7 +19,7 @@ const getStatus = (status: SettingStatusType) => {
   return <Component />;
 };
 
-const CommunityPage = () => {
+const CommunitySettingModal = () => {
   const { settingStatus } = useCommunityStore();
   const { setShowModal } = useModalStore();
 
@@ -28,38 +28,29 @@ const CommunityPage = () => {
   };
 
   return (
-    <BackgroundModal width={800} p={0}>
-      <SettingBox>
-        <Side>
-          <CommunitySettingBar />
-        </Side>
-        <Container>
-          <CancelIconWrapper onClick={closeModal}>
-            <CancelIcon />
-          </CancelIconWrapper>
-          {getStatus(settingStatus)}
-        </Container>
-      </SettingBox>
-    </BackgroundModal>
+    // <BackgroundModal width={0} p={0}>
+    <SettingBox>
+      <Side>
+        <CommunitySettingBar />
+      </Side>
+      <Container>
+        <CancelIconWrapper onClick={closeModal}>
+          <CancelIcon />
+        </CancelIconWrapper>
+        {getStatus(settingStatus)}
+      </Container>
+    </SettingBox>
+    // </BackgroundModal>
   );
 };
 
-const CancelIconWrapper = styled.div`
-  font-size: 5rem;
-  color: ${({ theme }) => theme.color["auth-desc"]};
-  cursor: pointer;
-  position: absolute;
-  right: 500px;
-  top: 25px;
-`;
-
 const SettingBox = styled.div`
-  width: 100%;
-  position: absolute;
+  /* width: 100%; */
+  position: fixed;
+  z-index: 4;
   top: 0;
-  right: 0;
-  bottom: 0;
   left: 0;
+
   display: flex;
   justify-content: center;
   overflow: hidden;
@@ -70,6 +61,15 @@ const SettingBox = styled.div`
   background-color: ${({ theme }) => theme.backgroundColor["voice-icon"]};
 `;
 
+const CancelIconWrapper = styled.div`
+  font-size: 5rem;
+  color: ${({ theme }) => theme.color["auth-desc"]};
+  cursor: pointer;
+  position: absolute;
+  right: 500px;
+  top: 25px;
+`;
+
 const Container = styled.div`
   position: relative;
   display: flex;
@@ -77,10 +77,11 @@ const Container = styled.div`
   align-items: flex-start;
   background-color: ${({ theme }) => theme.backgroundColor["tab3"]};
 `;
+
 const Side = styled.div`
   display: flex;
   flex: 1 1 auto;
   justify-content: flex-end;
 `;
 
-export default CommunityPage;
+export default CommunitySettingModal;
