+++ src/frontend/src/components/organisms/Tab2Footer.tsx
@@ -1,3 +1,4 @@
+import useDeviceStore from "@store/useDeviceStore";
 import useModalStore from "@store/useModalStore";
 import styled from "styled-components";
 import HeadsetIcon from "../atoms/Icons/HeadsetIcon";
@@ -10,6 +11,7 @@ import UserOnOffButton from "../molecules/Button/UserOnOffButton";
 
 const Tab2Footer = () => {
   const { setModalType, setShowModal } = useModalStore();
+  const { mic, sound, setMic, setSound } = useDeviceStore();
 
   const showSettingModal = () => {
     setModalType("userSetting");
@@ -17,33 +19,33 @@ const Tab2Footer = () => {
   };
 
   return (
-    <>
-      <Tab2FooterContainer>
-        <InfoContainer>
-          <UserInfoButton status="1" />
-        </InfoContainer>
-        <ButtonContainer>
-          <UserOnOffButton
-            text="음소거"
-            OnIcon={<MicIcon />}
-            OffIcon={<MicOffIcon />}
-            onClick={() => null}
-          />
-          <UserOnOffButton
-            text="헤드셋 음소거"
-            OnIcon={<HeadsetIcon />}
-            OffIcon={<HeadsetOffIcon />}
-            onClick={() => null}
-          />
-          <UserOnOffButton
-            text="사용자 설정"
-            OnIcon={<SettingsIcon />}
-            OffIcon={<SettingsIcon />}
-            onClick={showSettingModal}
-          />
-        </ButtonContainer>
-      </Tab2FooterContainer>
-    </>
+    <Tab2FooterContainer>
+      <InfoContainer>
+        <UserInfoButton status="1" />
+      </InfoContainer>
+      <ButtonContainer>
+        <UserOnOffButton
+          text="음소거"
+          on={mic}
+          OnIcon={<MicIcon />}
+          OffIcon={<MicOffIcon />}
+          onClick={() => setMic(!mic)}
+        />
+        <UserOnOffButton
+          text="헤드셋 음소거"
+          on={sound}
+          OnIcon={<HeadsetIcon />}
+          OffIcon={<HeadsetOffIcon />}
+          onClick={() => setSound(!sound)}
+        />
+        <UserOnOffButton
+          text="사용자 설정"
+          OnIcon={<SettingsIcon />}
+          OffIcon={<SettingsIcon />}
+          onClick={showSettingModal}
+        />
+      </ButtonContainer>
+    </Tab2FooterContainer>
   );
 };
