--- src/frontend/src/pages/CommunityPage.stories.tsx
@@ -1,8 +1,0 @@
-import CommunitySetting from "./CommunityPage";
-
-export default {
-  title: "Pages/CommunitySetting",
-  component: CommunitySetting,
-};
-
-export const CommunitySettingPage = () => <CommunitySetting />;
